# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: agora API
  description: |
    bon-soleil Holdings の全AIエージェントが共有する基盤プラットフォーム agora の API仕様書。
    
    - キャラクター情報・画像取得は認証不要
    - 画像生成・アップロードは `X-API-Key` ヘッダー必須
  version: "0.1.0"
  contact:
    name: bon-soleil Holdings
    url: https://agora.bon-soleil.com
servers:
  - url: https://agora.bon-soleil.com
    description: Production
  - url: http://localhost:8810
    description: Local (alice-hetzner / jasmine)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /casts/{id}/{file}:
    get:
      tags: [casts]
      summary: キャラクター画像配信
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: alice
        - name: file
          in: path
          required: true
          schema:
            type: string
          example: main.jpg
      responses:
        "200":
          description: 画像ファイル
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
...